Fixed a bunch of things related to ps vita
Former-commit-id: 409de3d668c7911ba2e4da11ebda7c1671d6ea7d
Former-commit-id: a37f85d07f01abd11bbf9498c85894dba2b149d4 [formerly bb68fb62116823c387fefae16b406782517a638d]
Former-commit-id: 966c09b66eb43474bbd868527cac5737ebff2203

+++ modules/d_std/source/core/stdc/config.d
@@ -5,6 +5,6 @@ version(CustomRuntimeTest) version = UseCustomRuntime;
 
 version(UseCustomRuntime)
 {
-    alias c_long = int;
-    alias c_ulong = uint;
+    alias c_long = ptrdiff_t;
+    alias c_ulong = size_t;
 }
